+++ docx-wasm/js/doc-defaults.ts
@@ -13,8 +13,8 @@ export class DocDefaults {
     return this;
   }
 
-  spacing(spacing: number) {
-    this.runProperty = { ...this.runProperty, spacing };
+  characterSpacing(characterSpacing: number) {
+    this.runProperty = { ...this.runProperty, characterSpacing };
     return this;
   }
 }