+++ docx-wasm/js/settings.ts
@@ -3,11 +3,17 @@ export type DocVar = {
   val: string;
 };
 
+export type CharacterSpacingValues =
+  | "doNotCompress"
+  | "compressPunctuation"
+  | "compressPunctuationAndJapaneseKana";
+
 export class Settings {
   _docId: string | null = null;
   _docVars: DocVar[] = [];
   _defaultTabStop = 840;
   _adjustLineHeightInTable = false;
+  _characterSpacingControl: CharacterSpacingValues | null = null;
 
   docId(id: string) {
     this._docId = id;
@@ -26,6 +32,11 @@ export class Settings {
 
   adjustLineHeightInTable() {
     this._adjustLineHeightInTable = true;
+    return this;
+  }
+
+  characterSpacingControl(t: CharacterSpacingValues) {
+    this._characterSpacingControl = t;
     return this;
   }
 }